requirements.txt
# pandas>=1.3.0
# numpy>=1.21.0
# faiss-cpu>=1.7.0
# sentence-transformers>=2.0.0
# beautifulsoup4>=4.9.0
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Vector search and similarity
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Traditional keyword search
scikit-learn>=1.3.0

# HTML parsing
beautifulsoup4>=4.12.0

# Sparse matrix operations
scipy>=1.10.0

# Deep learning backend (automatically installed with sentence-transformers)
torch>=2.0.0
transformers>=4.30.0

# HTTP requests with retry logic
requests>=2.28.0
urllib3>=1.26.0